cogstream-py/requirements.txt
numpy==1.19.4
opencv-python
websocket-client
